# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend REST API for golang nuryanto2121",
        "title": "Base Playtopia",
        "contact": {
            "name": "Nuryanto",
            "url": "https://www.linkedin.com/in/nuryanto-1b2721156/",
            "email": "nuryantofattih@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/account/fileupload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FileUpload"
                ],
                "summary": "File Upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "account image",
                        "name": "upload_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "path images",
                        "name": "path",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FileUpload"
                ],
                "summary": "Delete FileUpload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #pakai yang file_path ajah",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FileResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "errors": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "models.FileResponse": {
            "type": "object",
            "properties": {
                "file_name": {
                    "description": "ID       int    `json:\"id\"`",
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}